+++ Tasks/UnitySetup/UnitySetupV2/unity-editor-activation.ts
@@ -1,5 +1,5 @@
 import tl = require('azure-pipelines-task-lib/task');
-import { usernameInputVariableName, passwordInputVariableName, serialInputVariableName, activateLicenseInputVariableName } from './unity-setup';
+import { usernameInputVariableName, passwordInputVariableName, serialInputVariableName, activateLicenseInputVariableName } from './variables';
 import { getProjectUnityVersion } from './utilities';
 import path = require('path');
 import { UnityPathTools, Utilities } from '@dinomite-studios/unity-azure-pipelines-tasks-lib';
