+++ Tasks/UnitySetup/UnitySetupV2/unity-editor-install.ts
@@ -2,7 +2,7 @@ import tl = require('azure-pipelines-task-lib/task');
 import path = require('path');
 import { OS, Utilities } from '@dinomite-studios/unity-azure-pipelines-tasks-lib/';
 import { getProjectUnityVersion, getUnityHubExecutablePath } from './utilities';
-import { macOSArchitectureVariableName } from './unity-setup';
+import { macOSArchitectureVariableName, installEditorInputVariableName } from './variables';
 
 export class UnityEditorInstall {
     public static run(): number {
