+++ Tasks/UnitySetup/UnitySetupV2/unity-modules-install.ts
@@ -1,4 +1,4 @@
-import { androidChildModulesInputVariableName, androidModuleInputVariableName, iOSModuleInputVariableName, linuxIL2CPPModuleInputVariableName, linuxMonoModuleInputVariableName, macIL2CPPModuleInputVariableName, macMonoModuleInputVariableName, tvOSModuleInputVariableName, uwpModuleInputVariableName, visionOSModuleInputVariableName, webGLModuleInputVariableName, windowsModuleInputVariableName } from "./unity-setup";
+import { installEditorInputVariableName, androidChildModulesInputVariableName, androidModuleInputVariableName, iOSModuleInputVariableName, linuxIL2CPPModuleInputVariableName, linuxMonoModuleInputVariableName, macIL2CPPModuleInputVariableName, macMonoModuleInputVariableName, tvOSModuleInputVariableName, uwpModuleInputVariableName, visionOSModuleInputVariableName, webGLModuleInputVariableName, windowsModuleInputVariableName } from "./variables";
 import tl = require('azure-pipelines-task-lib/task');
 import { getProjectUnityVersion, getUnityHubExecutablePath } from "./utilities";
