+++ Tasks/UnitySetup/UnitySetupV2/unity-setup.ts
@@ -5,32 +5,6 @@ import { UnityModulesInstall } from './unity-modules-install';
 // import { UnityEditorConfiguration } from './unity-editor-configuration';
 import { UnityEditorActivation } from './unity-editor-activation';
 
-// Input variables
-export const versionSelectionModeVariableName = "versionSelectionMode";
-export const versionInputVariableName = 'version';
-export const revisionInputVariableName = 'revision';
-export const activateLicenseInputVariableName = 'activateLicense';
-export const usernameInputVariableName = 'username';
-export const passwordInputVariableName = 'password';
-export const serialInputVariableName = 'serial';
-export const deactivateSeatOnCompleteInputVariableName = 'deactivateSeatOnComplete';
-export const unityHubExecutableLocationVariableName = 'unityHubExecutableLocation';
-export const customUnityHubExecutableLocationVariableName = 'customUnityHubExecutableLocation';
-export const macOSArchitectureVariableName = 'macOSArchitecture';
-export const androidModuleInputVariableName = 'installAndroidModule';
-export const androidChildModulesInputVariableName = 'installAndroidChildModules';
-export const androidConfigureToolingInputVariableName = 'configureAndroidTooling';
-export const iOSModuleInputVariableName = 'installIOSModule';
-export const tvOSModuleInputVariableName = 'installTVOSModule';
-export const visionOSModuleInputVariableName = 'installVisionOSModule';
-export const linuxMonoModuleInputVariableName = 'installLinuxMonoModule';
-export const linuxIL2CPPModuleInputVariableName = 'installLinuxIL2CPPModule';
-export const macMonoModuleInputVariableName = 'installMacMonoModule';
-export const macIL2CPPModuleInputVariableName = 'installMacIL2CPPModule';
-export const windowsModuleInputVariableName = 'installWindowsIL2CPPModule';
-export const uwpModuleInputVariableName = 'installUWPModule';
-export const webGLModuleInputVariableName = 'installWebGLModule';
-
 function run() {
     try {
         // Configure localization.
